+++ decode_wav/src/lib.rs
@@ -137,7 +137,7 @@ impl Decoder for SymphoniaDecoder {
             params,
             num_frames,
             num_channels: num_channels as u16,
-            sample_rate: sample_rate.map(|s| s as f64),
+            sample_rate: sample_rate.map(|s| s as u32),
         };
 
         Ok((
